Remove preferences file
