--- qiskit/chemistry/preferences.py
@@ -1,69 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Copyright 2018 IBM.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# =============================================================================
-
-import os
-import json
-
-
-class Preferences(object):
-
-    _FILENAME = '.qiskit_chemistry'
-    _VERSION = '1.0'
-
-    def __init__(self):
-        """Create Preferences object."""
-        self._preferences = {
-            'version': Preferences._VERSION
-        }
-        self._logging_config_changed = False
-
-        home = os.path.expanduser("~")
-        self._filepath = os.path.join(home, Preferences._FILENAME)
-        try:
-            with open(self.filepath) as json_pref:
-                self._preferences = json.load(json_pref)
-                # remove old packages entry
-                if 'packages' in self._preferences:
-                    del self._preferences['packages']
-        except:
-            pass
-
-    @property
-    def filepath(self):
-        return self._filepath
-
-    def save(self):
-        if self._logging_config_changed:
-            with open(self.filepath, 'w') as fp:
-                json.dump(self._preferences, fp, sort_keys=True, indent=4)
-            self._logging_config_changed = False
-
-    def get_version(self):
-        if 'version' in self._preferences:
-            return self._preferences['version']
-
-        return None
-
-    def get_logging_config(self, default_value=None):
-        if 'logging_config' in self._preferences:
-            return self._preferences['logging_config']
-
-        return default_value
-
-    def set_logging_config(self, logging_config):
-        self._logging_config_changed = True
-        self._preferences['logging_config'] = logging_config
