+++ qiskit_chemistry_ui/__init__.py
@@ -14,3 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # =============================================================================
+
+from ._uipreferences import UIPreferences
+
+__all__ = ['UIPreferences']
